+++ xquare-design-system-android/src/main/res/values/attrs.xml
@@ -13,18 +13,18 @@
         <attr name="leadingSrcTint" format="color" />
         <attr name="trailingSrcTint" format="color" />
         <attr name="android:textStyle" format="reference" />
-        <attr name="android:textAppearance" format="reference"/>
+        <attr name="android:textAppearance" format="reference" />
         <attr name="android:textColor" format="reference|integer" />
         <attr name="android:fontFamily" format="reference" />
         <attr name="leadingSrcSize" format="dimension" />
-        <attr name="trailingSrcSize" format="dimension"/>
+        <attr name="trailingSrcSize" format="dimension" />
     </declare-styleable>
 
-    <declare-styleable name="XIconButton">
+    <declare-styleable name="IconXButton">
         <attr name="android:enabled" format="boolean" />
         <attr name="android:src" format="reference|integer" />
         <attr name="android:tint" format="color" />
-        <attr name="srcSize" format="dimension"/>
+        <attr name="srcSize" format="dimension" />
     </declare-styleable>
 
 </resources>
